MASSIVE REFACTOR: migrate integers to use cstdint types (where important for network compatibility)

+++ network.hpp
@@ -4,6 +4,7 @@
 #include <SFML/Network.hpp>
 
 #include <thread>
+#include <cstdint>
 
 class player_t;
 class world_t;
@@ -13,7 +14,7 @@ class Game;
 class Network {
 	private:
 	sf::UdpSocket socket;
-	unsigned short port;
+	uint16_t port;
 	sf::IpAddress host;
 	Game *game;
 
@@ -22,18 +23,18 @@ class Network {
 	float latency = 0; // measured in seconds
 
 	// measured in bytes - LAST_*X is reset per second
-	unsigned long long TX = 0, RX = 0;
-	unsigned long long DISP_TX = 0, DISP_RX = 0;
-	unsigned long long LAST_TX=0, LAST_RX = 0;
+	uint64_t TX = 0, RX = 0;
+	uint64_t DISP_TX = 0, DISP_RX = 0;
+	uint64_t LAST_TX=0, LAST_RX = 0;
 
-	Network(Game *game, sf::IpAddress addr, unsigned short port);
+	Network(Game *game, sf::IpAddress addr, uint16_t port);
 
 	bool connect(std::string username);
 	void disconnect();
 
 	void run();
-	player_t* getPlayerByID(unsigned short id);
-	zombie_t* getZombieByID(unsigned short id);
+	player_t* getPlayerByID(uint16_t id);
+	zombie_t* getZombieByID(uint16_t id);
 	void stop();
 	void send(sf::Packet packet);
 };
